% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bananarama-package.R
\docType{package}
\name{bananarama-package}
\alias{bananarama-package}
\title{bananarama: Generate Presentation Images Using Google Gemini}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Reproducibly generate slide images using Google Gemini. Define your images in a YAML configuration file with support for reference images and style defaults, eliminating the need to copy and paste prompts back and forth from a web UI. Images are generated in parallel for speed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hadley/bananarama}
  \item \url{https://hadley.github.io/bananarama/}
  \item Report bugs at \url{https://github.com/hadley/bananarama/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co}

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
